RAG_QDrant/requirements.txt
qdrant-client==1.12.1

langchain==0.3.12

langchain-huggingface==0.1.2

langchain-openai==0.2.12

sentence-transformers==3.3.1

python-dotenv==1.0.0

pdfplumber==0.5.28